+++ src/PointHandler.java
@@ -27,9 +27,9 @@ public class PointHandler implements Runnable {
         try {
             DataInputStream distream = new DataInputStream(socket.getInputStream());
 
-            float x = (float)distream.readDouble();
-            float y = (float)distream.readDouble();
-            float r = (float)distream.readDouble();
+            float x = (float)distream.readFloat();
+            float y = (float)distream.readFloat();
+            float r = (float)distream.readFloat();
 
             ServerArea serverArea = new ServerArea(r);
             Mark m = new Mark(x, y);
